refactor: reorganize Permissions import paths and enhance type imports
